requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --index-url=https://mirrors.aliyun.com/pypi/simple/ --output-file=requirements.txt requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple/

aiofiles==25.1.0
    # via ncatbot
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
apscheduler==3.11.2
    # via -r requirements.in
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via
    #   ncatbot
    #   qrcode
    #   tqdm
distro==1.9.0
    # via openai
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   ncatbot
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via ncatbot
jiter==0.13.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.10
    # via -r requirements.in
langchain-core==1.2.14
    # via
    #   langchain
    #   langchain-openai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.1.10
    # via -r requirements.in
langgraph==1.0.9
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.8
    # via langgraph
langgraph-sdk==0.3.6
    # via langgraph
langsmith==0.7.5
    # via langchain-core
lunardate==0.2.2
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
ncatbot==4.4.1.post1
    # via -r requirements.in
openai==2.21.0
    # via langchain-openai
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   langchain-core
    #   langsmith
    #   ncatbot
psutil==7.2.2
    # via ncatbot
pydantic==2.12.5
    # via
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
python-dotenv==1.2.1
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   langchain-core
    #   ncatbot
qrcode==8.2
    # via ncatbot
regex==2026.2.19
    # via tiktoken
requests==2.32.5
    # via
    #   langsmith
    #   ncatbot
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.3.3
    # via ncatbot
schedule==1.2.2
    # via ncatbot
sniffio==1.3.1
    # via openai
tenacity==9.1.4
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
toml==0.10.2
    # via ncatbot
tqdm==4.67.3
    # via
    #   ncatbot
    #   openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   langchain-core
    #   ncatbot
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via tzlocal
tzlocal==5.3.1
    # via apscheduler
urllib3==2.6.3
    # via
    #   ncatbot
    #   requests
uuid-utils==0.14.0
    # via
    #   langchain-core
    #   langsmith
websockets==16.0
    # via ncatbot
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
